Removes an extra class
