+++ backend/services/MoviesRepository.php
@@ -6,8 +6,10 @@ namespace Wtw\backend\services;
  * Interface for interacting with the remote API
  *
  * @param string $idData Information Identifier
+ *
+ * @return array
  */
 interface MoviesRepository
 {
-    public function getData(string $idData);
+    public function getData(string $imdbId): array;
 }
