+++ backend/services/OmdbRepository.php
@@ -2,6 +2,8 @@
 
 namespace Wtw\backend\services;
 
+use GuzzleHttp\Client;
+use Psr\Http\Message\ResponseInterface;
 use Wtw\backend\services\MoviesRepository;
 use Wtw\backend\services\HttpClient;
 
@@ -9,24 +11,33 @@ use Wtw\backend\services\HttpClient;
  * Repository for getting information about a movie by its ID from an external service
  * using the Remote Repository interface
  *
- * @param HttpClient $httpClient Http client for connecting to an external service
- * @param string     $idData     Movie ID
+ * @param string $apikey The user's api key for accessing the remote service
+ * @param string $idData Movie ID
  */
 class OmdbRepository implements MoviesRepository
 {
-    /**
-     * @var ClientInterface PSR compatible http client
-     */
-    private $client;
+    private string $apikey;
 
-    public function __construct(HttpClient $httpClient)
+    public function __construct(string $apikey)
     {
-        $this->client = $httpClient;
+        $this->apikey = $apikey;
     }
 
-    public function getData(string $idData): ?array
+    private function sendRequest(string $imdbId): ResponseInterface
     {
-        $response = $this->client->sendRequest($idData);
+        $client = new Client();
+
+        return $client->request('GET', 'http://www.omdbapi.com', [
+            'query' => [
+                'apikey' => $this->apikey,
+                'i' => $imdbId,
+            ]
+        ]);
+    }
+
+    public function getData(string $imdbId): array
+    {
+        $response = $this->sendRequest($imdbId);
 
         return json_decode($response->getBody()->getContents(), true);
     }
